+++ CHANGELOG.md
@@ -4,6 +4,8 @@ CHANGELOG
 ## HEAD (Unreleased)
 * Namespace names in .NET SDK are adjusted to PascalCase
 ([#38](https://github.com/pulumi/pulumi-f5bigip/pull/38)).
+* Upgrade to v1.1.1 of the BigIP Terraform Provider
+* Upgrade to pulumi-terraform-bridge v1.5.2
 
 ---